carousel update
Implemented the DRY rule by making the carousel component reusable.  I also stored the data in JSON files so it'll serve as a temporary "database" for the carousels
