+++ src/App.js
@@ -9,6 +9,7 @@ import Footer from './components/Footer';
 import LgNavbar from './components/LG-navbar';
 
 
+
 function App() {
   return (
     <div className="App">
