+++ src/Data/DiscountData.json
@@ -1,0 +1,100 @@
+[
+    {
+        "image": "images/event.png",
+        "name": "Jerry Mariot Hotel",
+        "rating": "4.5",
+        "category": "Hotel",
+        "duration": "All day"
+    },
+    {
+        "image": "images/event.png",
+        "name": "Mariot Hotel",
+        "rating": "4.5",
+        "category": "Lecture hall",
+        "duration": "All day"
+    },
+    {
+        "image": "images/event.png",
+        "name": "timmys house",
+        "rating": "4.5",
+        "category": "Lecture hall",
+        "duration": "All day"
+    },
+    {
+        "image": "images/event.png",
+        "name": "lorem Hotel",
+        "rating": "4.5",
+        "category": "Lodge",
+        "duration": "5 hours"
+    },
+    {
+        "image": "images/event.png",
+        "name": "Starrr",
+        "rating": "5.0",
+        "category": "Lodge",
+        "duration": "5 hours"
+    },
+    {
+        "image": "images/event.png",
+        "name": "Beach",
+        "rating": "3.0",
+        "category": "Lodge",
+        "duration": "7 hours"
+    },
+    {
+        "image": "images/event.png",
+        "name": "Ruffle",
+        "rating": "5.0",
+        "category": "Lodge",
+        "duration": "2 hours"
+    },
+    {
+        "image": "images/event.png",
+        "name": "Nurse hall",
+        "rating": "4.5",
+        "category": "Hotel",
+        "duration": "All day"
+    },
+    {
+        "image": "images/event.png",
+        "name": "Old caro",
+        "rating": "2.5",
+        "category": "Lecture hall",
+        "duration": "All day"
+    },
+    {
+        "image": "images/event.png",
+        "name": "Hilltop",
+        "rating": "1.5",
+        "category": "Lecture hall",
+        "duration": "All day"
+    },
+    {
+        "image": "images/event.png",
+        "name": "Odim",
+        "rating": "5.0",
+        "category": "Lodge",
+        "duration": "5 hours"
+    },
+    {
+        "image": "images/event.png",
+        "name": "Starrr",
+        "rating": "3.5",
+        "category": "Lodge",
+        "duration": "5 hours"
+    },
+    {
+        "image": "images/event.png",
+        "name": "Beach",
+        "rating": "3.0",
+        "category": "Lodge",
+        "duration": "7 hours"
+    },
+    {
+        "image": "images/event.png",
+        "name": "Odenugwe",
+        "rating": "5.0",
+        "category": "Lodge",
+        "duration": "2 hours"
+    }
+]
