+++ src/Data/EventData.json
@@ -1,0 +1,100 @@
+[
+    {   "id":"",
+        "image": "images/event.png",
+        "name": "stephanie's party",
+        "rating": "5.0",
+        "category": "Hotel",
+        "duration": "All day"
+    },
+    {   "id":"",
+        "image": "images/event.png",
+        "name": "lorem's meeting",
+        "rating": "2.5",
+        "category": "Lecture hall",
+        "duration": "All day"
+    },
+    {    "id":"",
+        "image": "images/place.png",
+        "name": "Healh walk",
+        "rating": "5.0",
+        "category": "Lecture hall",
+        "duration": "All day"
+    },
+    {   "id":"",
+        "image": "images/event.png",
+        "name": "Night of Glory",
+        "rating": "5.0",
+        "category": "Lodge",
+        "duration": "5 hours"
+    },
+    {   "id":"",
+        "image": "images/place.png",
+        "name": "Starrr",
+        "rating": "5.0",
+        "category": "Lodge",
+        "duration": "5 hours"
+    },
+    {   "id":"",
+        "image": "images/place.png",
+        "name": "Beach",
+        "rating": "5.0",
+        "category": "Lodge",
+        "duration": "7 hours"
+    },
+    {   "id":"",
+        "image": "images/place.png",
+        "name": "Ruffle",
+        "rating": "3.0",
+        "category": "Lodge",
+        "duration": "2 hours"
+    },
+    {   "id":"",
+        "image": "images/event.png",
+        "name": "Nurse hall",
+        "rating": "4.5",
+        "category": "Hotel",
+        "duration": "All day"
+    },
+    {   "id":"",
+        "image": "images/place.png",
+        "name": "Old caro",
+        "rating": "2.5",
+        "category": "Lecture hall",
+        "duration": "All day"
+    },
+    {   "id":"",
+        "image": "images/event.png",
+        "name": "Hilltop",
+        "rating": "3.5",
+        "category": "Lecture hall",
+        "duration": "All day"
+    },
+    {   "id":"",
+        "image": "images/place.png",
+        "name": "Odim",
+        "rating": "4.5",
+        "category": "Lodge",
+        "duration": "5 hours"
+    },
+    {   "id":"",
+        "image": "images/place.png",
+        "name": "Starrr",
+        "rating": "2.5",
+        "category": "Lodge",
+        "duration": "5 hours"
+    },
+    {   "id":"",
+        "image": "images/event.png",
+        "name": "Beach",
+        "rating": "3.0",
+        "category": "Lodge",
+        "duration": "7 hours"
+    },
+    {   "id":"",
+        "image": "images/place.png",
+        "name": "Odenugwe",
+        "rating": "1.5",
+        "category": "Lodge",
+        "duration": "2 hours"
+    }
+]
