+++ src/Data/PlaceData.json
@@ -1,0 +1,100 @@
+[
+    {
+        "image": "images/place.png",
+        "name": "vet hall",
+        "rating": "4.5",
+        "category": "Hotel",
+        "duration": "All day"
+    },
+    {
+        "image": "images/place.png",
+        "name": "Pharm hall",
+        "rating": "2.5",
+        "category": "Lecture hall",
+        "duration": "All day"
+    },
+    {
+        "image": "images/place.png",
+        "name": "Pharm hall",
+        "rating": "5.0",
+        "category": "Lecture hall",
+        "duration": "All day"
+    },
+    {
+        "image": "images/place.png",
+        "name": "Nice view",
+        "rating": "3.5",
+        "category": "Lodge",
+        "duration": "5 hours"
+    },
+    {
+        "image": "images/place.png",
+        "name": "Starrr",
+        "rating": "5.0",
+        "category": "Lodge",
+        "duration": "5 hours"
+    },
+    {
+        "image": "images/place.png",
+        "name": "Beach",
+        "rating": "3.0",
+        "category": "Lodge",
+        "duration": "7 hours"
+    },
+    {
+        "image": "images/place.png",
+        "name": "Ruffle",
+        "rating": "5.0",
+        "category": "Lodge",
+        "duration": "2 hours"
+    },
+    {
+        "image": "images/place.png",
+        "name": "Nurse hall",
+        "rating": "4.5",
+        "category": "Hotel",
+        "duration": "All day"
+    },
+    {
+        "image": "images/place.png",
+        "name": "Old caro",
+        "rating": "2.5",
+        "category": "Lecture hall",
+        "duration": "All day"
+    },
+    {
+        "image": "images/place.png",
+        "name": "Hilltop",
+        "rating": "5.0",
+        "category": "Lecture hall",
+        "duration": "All day"
+    },
+    {
+        "image": "images/place.png",
+        "name": "Odim",
+        "rating": "4.0",
+        "category": "Lodge",
+        "duration": "5 hours"
+    },
+    {
+        "image": "images/place.png",
+        "name": "Starrr",
+        "rating": "5.0",
+        "category": "Lodge",
+        "duration": "5 hours"
+    },
+    {
+        "image": "images/place.png",
+        "name": "Beach",
+        "rating": "3.0",
+        "category": "Lodge",
+        "duration": "7 hours"
+    },
+    {
+        "image": "images/place.png",
+        "name": "Odenugwe",
+        "rating": "5.0",
+        "category": "Lodge",
+        "duration": "2 hours"
+    }
+]
