+++ src/pages/Home.js
@@ -1,6 +1,5 @@
 import React from 'react'
 import Ad from '../components/Ad'
-import Carousel from '../components/Carousel'
 import Categories from '../components/Categories'
 import Places from '../components/Places'
 import Events from '../components/Events'
@@ -12,12 +11,10 @@ const index = () => {
     return (
         <div>
            <Ad text="Reserve online today, browse our wide assortment of rooms at different relaxation locations."/>
-           <Carousel />
            <Categories />
            <DiscountOffers/>
            <Events/>
            <Businesses />
-           
            <Places/>
            <Why />
         </div>
